// src/api/zzirit/models/TimeDealFetchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zzirit API
 * 찌릿 API 명세서
 *
 * The version of the OpenAPI document: v0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TimeDealFetchItem } from './TimeDealFetchItem';
import {
    TimeDealFetchItemFromJSON,
    TimeDealFetchItemFromJSONTyped,
    TimeDealFetchItemToJSON,
    TimeDealFetchItemToJSONTyped,
} from './TimeDealFetchItem';

/**
 * 
 * @export
 * @interface TimeDealFetchResponse
 */
export interface TimeDealFetchResponse {
    /**
     * 
     * @type {number}
     * @memberof TimeDealFetchResponse
     */
    timeDealId?: number;
    /**
     * 
     * @type {string}
     * @memberof TimeDealFetchResponse
     */
    timeDealName?: string;
    /**
     * 
     * @type {Date}
     * @memberof TimeDealFetchResponse
     */
    startTime?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TimeDealFetchResponse
     */
    endTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof TimeDealFetchResponse
     */
    status?: TimeDealFetchResponseStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof TimeDealFetchResponse
     */
    discountRatio?: number;
    /**
     * 
     * @type {Array<TimeDealFetchItem>}
     * @memberof TimeDealFetchResponse
     */
    items?: Array<TimeDealFetchItem>;
}


/**
 * @export
 */
export const TimeDealFetchResponseStatusEnum = {
    Scheduled: 'SCHEDULED',
    Ongoing: 'ONGOING',
    Ended: 'ENDED'
} as const;
export type TimeDealFetchResponseStatusEnum = typeof TimeDealFetchResponseStatusEnum[keyof typeof TimeDealFetchResponseStatusEnum];


/**
 * Check if a given object implements the TimeDealFetchResponse interface.
 */
export function instanceOfTimeDealFetchResponse(value: object): value is TimeDealFetchResponse {
    return true;
}

export function TimeDealFetchResponseFromJSON(json: any): TimeDealFetchResponse {
    return TimeDealFetchResponseFromJSONTyped(json, false);
}

export function TimeDealFetchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimeDealFetchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'timeDealId': json['timeDealId'] == null ? undefined : json['timeDealId'],
        'timeDealName': json['timeDealName'] == null ? undefined : json['timeDealName'],
        'startTime': json['startTime'] == null ? undefined : (new Date(json['startTime'])),
        'endTime': json['endTime'] == null ? undefined : (new Date(json['endTime'])),
        'status': json['status'] == null ? undefined : json['status'],
        'discountRatio': json['discountRatio'] == null ? undefined : json['discountRatio'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(TimeDealFetchItemFromJSON)),
    };
}

export function TimeDealFetchResponseToJSON(json: any): TimeDealFetchResponse {
    return TimeDealFetchResponseToJSONTyped(json, false);
}

export function TimeDealFetchResponseToJSONTyped(value?: TimeDealFetchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'timeDealId': value['timeDealId'],
        'timeDealName': value['timeDealName'],
        'startTime': value['startTime'] == null ? undefined : ((value['startTime']).toISOString()),
        'endTime': value['endTime'] == null ? undefined : ((value['endTime']).toISOString()),
        'status': value['status'],
        'discountRatio': value['discountRatio'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(TimeDealFetchItemToJSON)),
    };
}